Move GCP and Stackdriver code into a misk-gcp module
Also tidy up some of the build dependencies
