+++ build.gradle
@@ -13,6 +13,7 @@ ext {
           guice                : 'com.google.inject:guice:4.1.0',
           guiceMultibindings   : 'com.google.inject.extensions:guice-multibindings:4.1.0',
           guiceServlet         : 'com.google.inject.extensions:guice-servlet:4.1.0',
+          javaxInject          : 'javax.inject:javax.inject:1',
           jettyServer          : 'org.eclipse.jetty:jetty-server:9.4.8.v20171121',
           jettyServlet         : 'org.eclipse.jetty:jetty-servlet:9.4.8.v20171121',
           jettyWebsocketServlet: 'org.eclipse.jetty.websocket:websocket-servlet:9.4.8.v20171121',
@@ -32,7 +33,7 @@ ext {
           metricsGraphite      : 'io.dropwizard.metrics:metrics-graphite:4.0.1',
           metricsSignalFx      : 'com.signalfx.public:signalfx-codahale:0.0.41',
           metricsStackDriver   : 'com.google.apis:google-api-services-monitoring:v3-rev428-1.23.0',
-          gcpDatastore         : 'com.google.cloud:google-cloud-datastore:1.14.0',
+          gcpDatastore         : 'com.google.cloud:google-cloud-datastore:1.26.0',
           gcpKms               : 'com.google.apis:google-api-services-cloudkms:v1-rev30-1.23.0',
           okHttpMockWebServer  : 'com.squareup.okhttp3:mockwebserver:3.9.1',
           junitApi             : 'org.junit.jupiter:junit-jupiter-api:5.0.2',
