+++ misk-gcp-testing/gradle.properties
@@ -1,0 +1,4 @@
+POM_ARTIFACT_ID=misk-gcp-testing
+POM_NAME=misk-gcp-testing
+POM_DESCRIPTION=A Misk module for tesing code that uses GCP
+POM_PACKAGING=jar
