+++ misk-gcp-testing/src/main/kotlin/misk/cloud/gcp/storage/LocalStorageRpcTest.kt
