+++ misk-gcp/build.gradle
@@ -1,0 +1,43 @@
+import org.junit.platform.console.options.Details
+
+apply plugin: 'kotlin'
+apply plugin: 'org.junit.platform.gradle.plugin'
+
+compileKotlin {
+  kotlinOptions {
+    jvmTarget = "1.8"
+    allWarningsAsErrors = true
+  }
+}
+compileTestKotlin {
+  kotlinOptions {
+    jvmTarget = "1.8"
+    allWarningsAsErrors = true
+  }
+}
+
+sourceSets {
+  main.java.srcDirs += 'src/main/kotlin/'
+  test.java.srcDirs += 'src/test/kotlin/'
+}
+
+junitPlatform {
+  details Details.VERBOSE
+}
+
+dependencies {
+  compile dep.metricsStackDriver
+  compile(dep.gcpDatastore) {
+    exclude group: 'com.google.protobuf'
+    exclude group: 'com.google.api.grpc'
+    exclude group: 'io.grpc'
+  }
+  compile dep.gcpKms
+  compile dep.gcpCloudStorage
+  compile project(':misk')
+
+  testCompile project(':misk-testing')
+  testCompile project(':misk-gcp-testing')
+}
+
+apply from: rootProject.file('gradle/gradle-mvn-push.gradle')
