+++ misk-gcp/gradle.properties
@@ -1,0 +1,4 @@
+POM_ARTIFACT_ID=misk-gcp
+POM_NAME=misk-gcp
+POM_DESCRIPTION=Misk module for GCP
+POM_PACKAGING=jar
