+++ misk-gcp/src/main/kotlin/misk/cloud/gcp/storage/BaseCustomStorageRpc.kt
