+++ misk-hibernate-testing/gradle.properties
@@ -1,4 +1,4 @@
 POM_ARTIFACT_ID=misk-hibernate-testing
-POM_NAME=Misk
-POM_DESCRIPTION=An open source application container from Square
+POM_NAME=misk-hibernate-testing
+POM_DESCRIPTION=A Misk module for tesing code that uses Hibernate
 POM_PACKAGING=jar
