+++ misk-hibernate/gradle.properties
@@ -1,4 +1,4 @@
 POM_ARTIFACT_ID=misk-hibernate
-POM_NAME=Misk
-POM_DESCRIPTION=An open source application container from Square
+POM_NAME=misk-hibernate
+POM_DESCRIPTION=A Misk module for the Hibernate ORM
 POM_PACKAGING=jar
