+++ misk-testing/gradle.properties
@@ -1,4 +1,4 @@
 POM_ARTIFACT_ID=misk-testing
-POM_NAME=Misk
-POM_DESCRIPTION=An open source application container from Square
+POM_NAME=misk-testing
+POM_DESCRIPTION=A Misk module for testing code that uses Misk
 POM_PACKAGING=jar
