+++ misk/gradle.properties
@@ -1,4 +1,4 @@
 POM_ARTIFACT_ID=misk
-POM_NAME=Misk
+POM_NAME=misk
 POM_DESCRIPTION=An open source application container from Square
 POM_PACKAGING=jar
