--- misk/src/main/kotlin/misk/metrics/MetricsConfig.kt
@@ -1,7 +1,0 @@
-package misk.metrics
-
-import misk.config.Config
-import misk.metrics.backends.MetricsBackendConfig
-
-data class MetricsConfig(val backends: MetricsBackendConfig) : Config
-
