+++ samples/exemplar/gradle.properties
@@ -1,4 +1,4 @@
 POM_ARTIFACT_ID=exemplar
-POM_NAME=Misk Exemplar Project
-POM_DESCRIPTION=Example project using Misk framework
+POM_NAME=exemplar
+POM_DESCRIPTION=A Misk sample project
 POM_PACKAGING=jar
