+++ samples/exemplarchat/gradle.properties
@@ -1,4 +1,4 @@
 POM_ARTIFACT_ID=chat
-POM_NAME=Misk Chat Project
+POM_NAME=chat
 POM_DESCRIPTION=Chat using Misk
 POM_PACKAGING=jar
