+++ samples/urlshortener/gradle.properties
@@ -1,4 +1,4 @@
 POM_ARTIFACT_ID=urlshortener
 POM_NAME=urlshortener
-POM_DESCRIPTION=A Misk Sample project
+POM_DESCRIPTION=A Misk sample project
 POM_PACKAGING=jar
